+++ tests/unit_tests/core/test_exceptions.py
