+++ tests/unit_tests/core/test_utils.py
